requirements.txt
chainlit==1.3.2
langchain==0.3.11
langchain_community==0.3.11
langchain_core==0.3.24
langchain_experimental==0.3.3
langchain_google_vertexai==2.0.8
matplotlib==3.9.3
numpy<=2.0
pandas==2.2.3
python-dotenv==1.0.1
seaborn==0.13.2
